requirements.txt
altgraph==0.17.5
greenlet==3.3.2
packaging==25.0
playwright==1.58.0
pyee==13.0.1
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.10
python-dotenv==1.2.1
setuptools==80.9.0
typing_extensions==4.15.0
